retailer_tracker/retailer_tracker/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-utils-six==2.0
mssql-django==1.3
phonenumbers==8.13.18
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3